Register event page route fixed
